services/script-gen/requirements.txt
# services/script-gen/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Used inside Docker (Python 3.9, Linux).
# Do NOT run this on Windows with Python 3.14 — camel-tools won't build.
# For local Windows development, use requirements-dev.txt instead.
# ─────────────────────────────────────────────────────────────────────────────

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# HTTP client (calls GLM API)
httpx==0.27.2

# Data validation
pydantic==2.9.2

# Caching
redis==5.2.0

# Environment variables
python-dotenv==1.0.1

# ── AI / ML ────────────────────────────────────────────────────────────────
# PyTorch CPU-only build (saves 2GB vs full CUDA build for the script-gen service)
# If you add GPU support later, remove +cpu and install the CUDA build
# The index URL is now managed in the Dockerfile for better reliability.
torch==2.1.0+cpu

# HuggingFace transformers for diacritization model
transformers==4.40.2
sentencepiece==0.2.0      # required by Arabic BERT tokenizer
protobuf==4.25.3          # required by sentencepiece

# camel-tools (Arabic NLP)
# Requires cmake + g++ build tools — installed in Dockerfile before this line
camel-tools==1.4.1

# ── Logging & monitoring ───────────────────────────────────────────────────
structlog==24.4.0

# ── Testing ────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
# httpx version is already defined above